allow shutdown while migrating
git-svn-id: file:///home/toad/git-migration/temprepository/branches/saltedhashstore/freenet@21020 67a373e5-eb02-0410-a15c-ee090a768436

+++ src/freenet/store/saltedhash/SaltedHashFreenetStore.java
@@ -1496,7 +1496,7 @@ public class SaltedHashFreenetStore implements FreenetStore {
 
 				byte[] b = new byte[(int) entryTotalLength];
 
-				while (true) {
+				while (!shutdown) {
 					int status = storeRAF.read(b);
 					if (status != entryTotalLength)
 						break;
@@ -1519,7 +1519,8 @@ public class SaltedHashFreenetStore implements FreenetStore {
 				}
 			} catch (IOException ioe) {
 			}
-			storeFiles.delete();	
+			if (!shutdown)
+				storeFiles.delete();	
 		}
 
 		System.out.println(name + ": old salt hash-->new salt hash migrated: " + c + " keys(done)");
